Group resource constraint calculations
Computes group-resource requirements on the cluster by taking
user-provided constraints into account. Jobs that could be run in
parallel will now be run in series if not enough resources (e.g cores or
memory) are available. Runtime is now treated as a special resource, and
will be accurately calculated for parallel jobs via max() or for series
jobs via sum()

Pipe groups are treated as a single unit when calculating resources. Runtime is
summed over pipe groups instead of "maxed". Cores are merged via max()