Fix sorting for pipe job scheduling.
The toposorting within GroupJob was updated to put all pipe jobs
at the same toposort level so that accurate resource requirements
could be calculated. This broke the scheduler, however, which
relied on pipe jobs within the same group being in an hierarchical
order. This adds an additional toposort to GroupJob.__iter__
specifically for pipe jobs.

This also refactors the toposorting code to a less complex form