Add resource constraints to cluster job submission
Include the --resources flag on the snakemake call made in the child
job.

Does not send any runtime information, as this falsly constrains jobs
within the cluster.

Cores is still set to "all" for regular cluster submission, leaving the
cluster to manage how many cores the job is entitled to